# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/drivers": {
      "get": {
        "description": "Returns drivers list. Driver role is allowed so driver portal profile page can load its own driver data.",
        "operationId": "DriversController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search driver by name, email, phone, vehicle, or license number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all drivers",
        "tags": [
          "Drivers"
        ]
      },
      "post": {
        "description": "Only superuser and fleet manager can create drivers.",
        "operationId": "DriversController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add a new driver",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/drivers/{id}": {
      "get": {
        "description": "Returns one driver profile. Driver role is allowed for driver portal profile view.",
        "operationId": "DriversController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get driver by ID",
        "tags": [
          "Drivers"
        ]
      },
      "put": {
        "description": "Superuser and fleet manager can update driver data. Driver role is allowed so driver portal profile page can save its own profile.",
        "operationId": "DriversController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update driver info",
        "tags": [
          "Drivers"
        ]
      },
      "delete": {
        "description": "Only superuser and fleet manager can delete drivers.",
        "operationId": "DriversController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a driver",
        "tags": [
          "Drivers"
        ]
      }
    },
...
    "/api/documents": {
      "get": {
        "operationId": "DocumentsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all documents",
        "tags": [
          "Documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a document",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get document by ID",
        "tags": [
          "Documents"
        ]
      },
      "put": {
        "operationId": "DocumentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a document",
        "tags": [
          "Documents"
        ]
      },
      "delete": {
        "operationId": "DocumentsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a document",
        "tags": [
          "Documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "DeliverSync API",
    "description": "REST API for the DeliverSync delivery management platform. Role-based access is controlled via the `x-user-role` request header. Supported roles: superuser, fleet-manager, business-client, driver.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateDriverDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Raghav Reddy"
          },
          "email": {
            "type": "string",
            "example": "driver@deliverysync.com"
          },
          "phone": {
            "type": "string",
            "example": "+91 9440011223"
          },
          "licenseNumber": {
            "type": "string",
            "example": "TN03201800034521"
          },
          "vehicle": {
            "type": "string",
            "example": "TN09AB1234"
          },
          "vehicleType": {
            "type": "string",
            "example": "Tempo"
          },
          "status": {
            "type": "string",
            "example": "Available",
            "enum": [
              "Available",
              "Active",
              "Busy",
              "Offline",
              "Inactive"
            ]
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "UpdateDriverDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Raghav Reddy"
          },
          "email": {
            "type": "string",
            "example": "driver@deliverysync.com"
          },
          "phone": {
            "type": "string",
            "example": "+91 9440011223"
          },
          "licenseNumber": {
            "type": "string",
            "example": "TN03201800034521"
          },
          "vehicle": {
            "type": "string",
            "example": "TN09AB1234"
          },
          "vehicleType": {
            "type": "string",
            "example": "Tempo"
          },
          "status": {
            "type": "string",
            "example": "Available",
            "enum": [
              "Available",
              "Active",
              "Busy",
              "Offline",
              "Inactive"
            ]
          }
        }
      },
...
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "ownerType": {
            "type": "string",
            "example": "vehicle",
            "enum": [
              "vehicle",
              "driver"
            ]
          },
          "vehicle": {
            "type": "string",
            "example": "TN-09-AB-2345"
          },
          "driver": {
            "type": "string",
            "example": "Rajesh Kumar"
          },
          "documentType": {
            "type": "string",
            "example": "Registration"
          },
          "issueDate": {
            "type": "string",
            "example": "2025-04-01"
          },
          "expiryDate": {
            "type": "string",
            "example": "2026-03-31"
          },
          "status": {
            "type": "string",
            "example": "Valid",
            "enum": [
              "Valid",
              "Expiring Soon",
              "Expired"
            ]
          }
        },
        "required": [
          "ownerType",
          "documentType"
        ]
      },
      "UpdateDocumentDto": {
        "type": "object",
        "properties": {
          "ownerType": {
            "type": "string",
            "example": "vehicle",
            "enum": [
              "vehicle",
              "driver"
            ]
          },
          "vehicle": {
            "type": "string",
            "example": "TN-09-AB-2345"
          },
          "driver": {
            "type": "string",
            "example": "Rajesh Kumar"
          },
          "documentType": {
            "type": "string",
            "example": "Insurance"
          },
          "issueDate": {
            "type": "string",
            "example": "2025-04-01"
          },
          "expiryDate": {
            "type": "string",
            "example": "2026-03-31"
          },
          "status": {
            "type": "string",
            "example": "Valid",
            "enum": [
              "Valid",
              "Expiring Soon",
              "Expired"
            ]
          }
        }
      },
...
    }
  }
}